+++ activerecord/test/cases/calculations_test.rb
@@ -389,8 +389,10 @@ class CalculationsTest < ActiveRecord::TestCase
         Account.where("credit_limit > 50").from('accounts').sum(:credit_limit)
   end
 
-  def test_sum_array_compatibility
-    assert_equal Account.sum(:credit_limit), Account.sum(&:credit_limit)
+  def test_sum_array_compatibility_deprecation
+    assert_deprecated do
+      assert_equal Account.sum(:credit_limit), Account.sum(&:credit_limit)
+    end
   end
 
   def test_average_with_from_option